requirements.txt
streamlitstreamlit==1.26.0
pillow==10.0.1
pandas==2.1.1
numpy==1.26.0
